+++ .npmignore
@@ -17,6 +17,7 @@ logs
 *.log
 npm-debug.log*
 .pnpm-debug.log*
+.pnpm-store
 
 # Compiled binary addons (https://nodejs.org/api/addons.html)
 build/Release
@@ -24,8 +25,8 @@ build/Release
 # Dependency directories
 node_modules/
 
-# generate output
-dist
+# Typescript
+tsconfig.json
 
 # TypeScript cache
 *.tsbuildinfo
