--- src/Seeds.js
@@ -1,1 +1,0 @@
-// TODO: Write seeders
