--- src/lib/IDBModel.js
@@ -1,1 +1,0 @@
-// IDBModel
